coffee/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
multidict==6.0.4
packaging==23.2
Pillow==10.1.0
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.0.6
yarl==1.9.2
